+++ src/simulations/machines_and_tapes_ex2/layout.tsx
@@ -1,38 +1,14 @@
 import p5 from "p5"
 import React, { CSSProperties } from "react"
 import ReactDOM from "react-dom"
-import { Breadcrumbs } from "../../react-components/breadcrumbs"
-import { ScreenShotButton } from "../../react-components/screenshot_button"
+import { DetailPage } from "../../react-components/detail_page"
 import { main, getTimestamp } from "./source"
 
 const App = () => {
-  const bodyStyle: CSSProperties = {
-    display: "table", // horizontal center
-    margin: "auto",
-    marginTop: "4rem",
-    border: "1px solid rgba(212, 214, 245, 1.0)",
-    borderRadius: "6px",
-    overflow: "hidden", // clip border-radius
-    position: "relative", // to make sure overflow works
-  }
-  const toolbarStyle: CSSProperties = {
-  }
-  const screenshotButtonStyle: CSSProperties = {
-    margin: "16px",
-  }
-
   return (
-    <div>
-      <Breadcrumbs/>
-      <div style={bodyStyle}>
-        <div id="canvas-parent"></div>
-        <div style={toolbarStyle}>
-          <div style={screenshotButtonStyle}>
-            <ScreenShotButton getTimestamp={() => getTimestamp()} />
-          </div>
-        </div>
-      </div>
-    </div>
+    <DetailPage getTimestamp={() => getTimestamp()}>
+      <div id="canvas-parent"></div>
+    </DetailPage>
   )
 }
